# backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Sthira /api/v3 boundary
  version: 3.0.0
  description: >-
    Contract for the Go backend. Documents the implemented health and
    middle-model validation boundary plus the P4 citizen destination/stay slice
    (sessions, place resolution, eligible-destination query, reservations and
    stay events). Package import/review/publish and speech endpoints are later
    phases and remain absent. All operational facts carry source provenance,
    evidence class and version; unknown is distinct from empty, zero or current.
    The P4 addition is deliberate: the P1 slice deferred these endpoints to a
    later phase; P4 implements the citizen stay flows.
servers:
  - url: /api/v3
...
paths:
...
  /voice/commands:
    post:
      tags: [voice]
      summary: Validate a constrained middle-model proposal
      description: >-
        Accepts a typed chat/transcript command carrying a middle-model
        proposal and independently validates it against the current server
        snapshot and active-context IDs. Performs no write, call or capacity
        mutation. Request bodies are bounded in size and depth; duplicate JSON
        keys, unknown fields, trailing data and prohibited actions are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VoiceCommandRequest' }
      responses:
        '200': { $ref: '#/components/responses/DataEnvelope' }
        '400': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/ErrorEnvelope' }
        '415': { $ref: '#/components/responses/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/ErrorEnvelope' }
  /sessions:
    post:
      tags: [stay]
      summary: Issue a citizen session
      description: >-
        Account-free citizen session issuance (R22): no civil identity is
        required. Returns an unguessable bearer token once; only its SHA-256
        hash is stored. Sessions expire and can be revoked. Public.
      responses:
        '201': { $ref: '#/components/responses/DataEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /places/resolve:
    post:
      tags: [stay]
      summary: Resolve a place by alias or admin ID
      description: >-
        Jurisdiction-aware resolution from known aliases/admin IDs. Multiple
        candidates yield AMBIGUOUS_PLACE with candidate IDs; unknown yields
        NOT_FOUND. No LLM geocode. Public.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jurisdiction, query]
              properties:
                jurisdiction: { type: string }
                query: { type: string }
      responses:
        '200': { $ref: '#/components/responses/DataEnvelope' }
        '404': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '409': { $ref: '#/components/responses/ErrorEnvelope' }
        '415': { $ref: '#/components/responses/ErrorEnvelope' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /guidance/query:
    post:
      tags: [stay]
      summary: Read-only eligible destination query
      description: >-
        Read-only preview of eligible destinations for a party and half-open
        date range. Unknown capacity is reported as unknown and never promised.
        The operational route gate is closed while route authority (O05) is
        open; no destination is route-verified for operational use. Performs no
        write. Public.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jurisdiction, package_id, party_size, start_date, end_date]
              properties:
                jurisdiction: { type: string }
                package_id: { type: string }
                party_size: { type: integer, minimum: 1 }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/DataEnvelope' }
        '400': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '415': { $ref: '#/components/responses/ErrorEnvelope' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /reservations:
    post:
      tags: [stay]
      summary: Create an explicit reservation
      description: >-
        Explicit confirmed choice, separate from the read-only preview. Binds
        facility+route+source snapshot+party size and revalidates the source
        snapshot version at commit (STALE_VERSION on drift). Idempotency-keyed;
        a capacity conflict is a recoverable CAPACITY_CONFLICT and never
        silently substitutes another destination. Requires a citizen session.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [facility_id, package_id, party_size, start_date, end_date, idempotency_key, snapshot_version]
              properties:
                facility_id: { type: string }
                package_id: { type: string }
                route_id: { type: string }
                party_size: { type: integer, minimum: 1 }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                snapshot_version:
                  type: integer
                  description: >-
                    Authoritative source snapshot version client confirmed its
                    choice against; revalidated at commit under the locked
                    snapshot. Required and must be > 0. Rejected with
                    INVALID_VALUE (400) if zero or missing. Returns
                    STALE_VERSION (409) if the locked package snapshot differs.
      responses:
        '201': { $ref: '#/components/responses/DataEnvelope' }
        '400': { $ref: '#/components/responses/ErrorEnvelope' }
        '401': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '409': { $ref: '#/components/responses/ErrorEnvelope' }
        '415': { $ref: '#/components/responses/ErrorEnvelope' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /reservations/{id}:
    get:
      tags: [stay]
      summary: Read own reservation/stay (restart recovery)
      description: >-
        Authenticated owner-only read of a reservation/stay by stay or
        reservation ID; the authenticated read path to recover state after a
        restart. Knowing the ID is not authorization (R22). Private, no-store.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/DataEnvelope' }
        '401': { $ref: '#/components/responses/ErrorEnvelope' }
        '403': { $ref: '#/components/responses/ErrorEnvelope' }
        '404': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /reservations/{id}/events:
    post:
      tags: [stay]
      summary: Apply a typed stay event
      description: >-
        Typed event ARRIVE/CANCEL/DEPART/EXTEND/TRANSFER on an owned stay.
        Arrival requires explicit citizen confirmation (R14). Idempotency-keyed;
        capacity conservation holds across every event. Requires the owning
        citizen session.
        For TRANSFER, the response includes new_stay_id and new_reservation_id
        for the replacement stay (both on first success and on idempotent replay).
        The authenticated GET /reservations/{new_stay_id} provides the recovery
        path after restart or a lost response.
        snapshot_version is REQUIRED for EXTEND and TRANSFER (those are
        commitments that revalidate against the locked snapshot). For ARRIVE,
        CANCEL, DEPART the snapshot_version may be omitted because no
        commitment revalidation is required.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, idempotency_key]
              properties:
                type: { type: string, enum: [ARRIVE, CANCEL, DEPART, EXTEND, TRANSFER] }
                idempotency_key: { type: string }
                new_end_date: { type: string, format: date }
                new_facility_id: { type: string }
                new_route_id: { type: string }
                snapshot_version:
                  type: integer
                  description: >-
                    Required for EXTEND and TRANSFER (snapshot_version > 0).
                    Rejected with INVALID_VALUE if zero or missing for those
                    commitment types.
      responses:
        '200': { $ref: '#/components/responses/DataEnvelope' }
        '400': { $ref: '#/components/responses/ErrorEnvelope' }
        '401': { $ref: '#/components/responses/ErrorEnvelope' }
        '403': { $ref: '#/components/responses/ErrorEnvelope' }
        '404': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '409': { $ref: '#/components/responses/ErrorEnvelope' }
        '415': { $ref: '#/components/responses/ErrorEnvelope' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /operations/sessions:
    post:
      tags: [operations]
      summary: Issue an operator session
      description: >-
        Issues a jurisdiction-scoped OPERATOR session. Operator authority comes
        from a server-verified identity plus a second factor (MFA) established at
        a trusted authentication boundary, NOT from any request-body flag (R22).
        The request carries no jurisdiction or MFA field: the permitted
        jurisdiction is derived from a server-controlled grant bound to the
        verified subject. With no trusted verifier configured the endpoint fails
        closed (503); without a grant it returns 403. The bearer token is shown
        once; only its hash is stored.
      responses:
        '201': { $ref: '#/components/responses/DataEnvelope' }
        '401': { $ref: '#/components/responses/ErrorEnvelope' }
        '403': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /operations/sources/{id}/transitions:
    post:
      tags: [operations]
      summary: Publish, suspend or revoke a source
      description: >-
        Moves a source to OPERATIONAL (publish), SUSPENDED or RETIRED (revoke).
        Requires an OPERATOR session with verified MFA. Jurisdiction-scoped: the
        source must hold a currently-valid authorization in the operator's own
        jurisdiction, else FORBIDDEN. Idempotency-keyed; the transition and its
        audit event commit atomically.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target, idempotency_key]
              properties:
                target: { type: string, enum: [OPERATIONAL, SUSPENDED, RETIRED] }
                idempotency_key: { type: string }
                reason: { type: string }
      responses:
        '200': { $ref: '#/components/responses/DataEnvelope' }
        '400': { $ref: '#/components/responses/ErrorEnvelope' }
        '401': { $ref: '#/components/responses/ErrorEnvelope' }
        '403': { $ref: '#/components/responses/ErrorEnvelope' }
        '404': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '409': { $ref: '#/components/responses/ErrorEnvelope' }
        '415': { $ref: '#/components/responses/ErrorEnvelope' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
  /operations/sources/{id}/quarantine:
    post:
      tags: [operations]
      summary: Quarantine a source's evidence
      description: >-
        Moves a source from any active state to QUARANTINED, marking its evidence
        as not-to-be-used pending review. Distinct from SUSPENDED (a temporary
        operational pause). Requires an OPERATOR session with verified MFA.
        Jurisdiction-scoped via the source's current authorization. Quarantined
        evidence is excluded from the resolved operational context and from new
        reservation commit revalidation; existing stays are NOT cancelled or
        released. Quarantine is terminal: release requires a fresh source.
        Idempotency-keyed to the target source; version-guarded; audit-attributed.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idempotency_key, reason]
              properties:
                idempotency_key: { type: string }
                reason: { type: string }
      responses:
        '200': { $ref: '#/components/responses/DataEnvelope' }
        '400': { $ref: '#/components/responses/ErrorEnvelope' }
        '401': { $ref: '#/components/responses/ErrorEnvelope' }
        '403': { $ref: '#/components/responses/ErrorEnvelope' }
        '404': { $ref: '#/components/responses/ErrorEnvelope' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '409': { $ref: '#/components/responses/ErrorEnvelope' }
        '415': { $ref: '#/components/responses/ErrorEnvelope' }
        '503': { $ref: '#/components/responses/ErrorEnvelope' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    DataEnvelope:
      description: Success envelope; exactly `data` is populated.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    ErrorEnvelope:
      description: Error envelope; exactly `errors` is populated.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    MethodNotAllowed:
      description: Method not allowed; `Allow` header names the permitted method.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
  schemas:
    Envelope:
      type: object
      required: [request_id, schema_version, generated_at, data_version, source_status]
      properties:
        request_id: { type: string }
        schema_version: { type: string, enum: ['3.0'] }
        generated_at: { type: string, format: date-time, description: UTC }
        data_version: { type: string }
        source_status:
          type: string
          enum: [CURRENT, STALE, EXPIRED, UNAVAILABLE, UNKNOWN]
        data: { description: Present on success only. }
        errors:
          type: array
          items: { $ref: '#/components/schemas/APIError' }
          description: Present on failure only.
...
    VoiceCommandRequest:
      type: object
      required: [request_id, data_version, jurisdiction, proposal]
      additionalProperties: false
      properties:
        request_id: { type: string }
        data_version: { type: string }
        jurisdiction:
          type: string
          description: >-
            Untrusted lookup input selecting WHICH jurisdiction's authoritative
            snapshot to resolve. Never proof of authorization or snapshot
            contents. Required; an unknown/empty jurisdiction resolves nothing
            (fail closed) rather than substituting another jurisdiction.
        proposal: { $ref: '#/components/schemas/ModelOutput' }
...